[fix] Setup page scroll view padding
The commit adjusts the padding for the `SingleChildScrollView` within the `SetupPage`.
The `padding` was changed from `EdgeInsets.all(10)` to `EdgeInsets.only(left: 15, right: 15, top: 60, bottom: 60)`.
This change modifies the spacing around the content within the scrollable area of the setup page.
The `Center` widget wrapping the `SingleChildScrollView` was removed as part of this change.